# app/api-node/swagger.yaml
openapi: 3.0.0
info:
    description: 'BattleBots API'
    license:
        name: MIT
    title: BattleBots
    version: 0.0.1
paths:
...
    '/api/arena/{arenaId}/bot/{botId}':
        put:
            operationId: ArenaLinkBot
            responses:
                '200':
                    description: 'link bot to arena'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIArenaResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIArenaResource'
                '404':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIArenaResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: arenaId
                    required: true
                    schema:
                        type: number
                        format: double
                -
                    description: ""
                    in: path
                    name: botId
                    required: true
                    schema:
                        type: number
                        format: double
    /api/bots/:
        get:
            operationId: BotsList
            responses:
                '200':
                    description: 'bot list'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '204':
                    description: 'No content'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
        post:
            operationId: BotsInsert
            responses:
                '201':
                    description: 'bot inserted'
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            requestBody:
                $ref: '#/components/requestBodies/IBotsResource'
        put:
            operationId: BotsUpdate
            responses:
                '200':
                    description: 'bot updated'
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            requestBody:
                $ref: '#/components/requestBodies/IBotsResource'
    '/api/bots/{id}':
        delete:
            operationId: BotsDelete
            responses:
                '200':
                    description: 'bot deleted'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '404':
                    description: 'bot not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
        get:
            operationId: BotsDetails
            responses:
                '200':
                    description: 'bot details'
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
                '404':
                    description: 'bot not found'
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIStreamResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
...
servers:
    -
        url: /
components:
    requestBodies:
...
        IBotsResource:
            content:
                'application/json; charset=UTF-8':
                    schema:
                        $ref: '#/components/schemas/IBotsResource'
            required: true
...
    securitySchemes:
        Bearer:
            type: http
            scheme: bearer
    schemas:
        IArenaResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                arena_name:
                    type: string
                    description: ""
                available:
                    type: number
                    format: double
                    description: ""
                bots:
                    type: array
                    items:
                        $ref: '#/components/schemas/IBotsResource'
                    description: ""
            type: object
        IBotsResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                botIp:
                    type: string
                    description: ""
                running:
                    type: number
                    format: double
                    description: ""
                taken:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                speed:
                    type: number
                    format: double
                    description: ""
                damage:
                    type: number
                    format: double
                    description: ""
                fireRate:
                    type: number
                    format: double
                    description: ""
                armor:
                    type: number
                    format: double
                    description: ""
                streams:
                    type: array
                    items:
                        $ref: '#/components/schemas/IStreamResource'
                    description: ""
                games:
                    type: array
                    items:
                        $ref: '#/components/schemas/IGameResource'
                    description: ""
                gameProfile:
                    $ref: '#/components/schemas/IGameProfileResource'
            type: object
            required:
                - id
                - botIp
                - running
                - taken
                - name
                - speed
                - damage
                - fireRate
                - armor
        IStreamResource:
            description: ""
            properties:
                game:
                    $ref: '#/components/schemas/IGameResource'
                robot:
                    $ref: '#/components/schemas/IBotsResource'
                kinesisUrl:
                    type: string
                    description: ""
                s3Url:
                    type: string
                    description: ""
                private:
                    type: number
                    format: double
                    description: ""
                running:
                    type: number
                    format: double
                    description: ""
                duration:
                    type: number
                    format: double
                    description: ""
                encodage:
                    type: string
                    description: ""
                id:
                    type: number
                    format: double
                    description: ""
            type: object
            required:
                - kinesisUrl
                - s3Url
                - private
                - running
                - duration
                - encodage
        IGameResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                arena:
                    $ref: '#/components/schemas/IArenaResource'
                bots:
                    type: array
                    items:
                        $ref: '#/components/schemas/IBotsResource'
                    description: ""
                streams:
                    type: array
                    items:
                        $ref: '#/components/schemas/IStreamResource'
                    description: ""
                status:
                    enum:
                        - CREATED
                        - STARTED
                        - STOPPED
                        - ENDED
                    type: string
                    description: ""
                createdAt:
                    type: string
                    description: ""
                updatedAt:
                    type: string
                    description: ""
                token:
                    type: string
                    description: ""
            type: object
            required:
                - name
                - status
        IGameProfileResource:
            description: ""
            properties:
                total_points:
                    type: number
                    format: double
                    description: ""
                id:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                user:
                    $ref: '#/components/schemas/IUserResource'
            type: object
        IUserResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                firstname:
                    type: string
                    description: ""
                lastname:
                    type: string
                    description: ""
                pseudo:
                    type: string
                    description: ""
                email:
                    type: string
                    description: ""
                roles:
                    type: string
                    description: ""
                address:
                    type: string
                    description: ""
                createdAt:
                    type: string
                    description: ""
                updatedAt:
                    type: string
                    description: ""
                password:
                    type: string
                    description: ""
                gameProfile:
                    type: array
                    items:
                        $ref: '#/components/schemas/IGameProfileResource'
                    description: ""
            type: object
            required:
                - pseudo
                - email
                - roles
        HttpResponseModelIArenaResource:
            description: ""
            properties:
                message:
                    type: string
                    description: ""
                httpCode:
                    type: number
                    format: double
                    description: ""
                data:
                    $ref: '#/components/schemas/IArenaResource'
            type: object
...
        HttpResponseModelIStreamResource:
            description: ""
            properties:
                message:
                    type: string
                    description: ""
                httpCode:
                    type: number
                    format: double
                    description: ""
                data:
                    $ref: '#/components/schemas/IStreamResource'
            type: object
...